DeepSeekWeatherMPC/requirements.txt
requests==2.31.0
cryptography==41.0.5
tqdm==4.66.1
python-dotenv==1.0.0
pycryptodome==3.19.0